+++ src/snowflake/cli/plugins/spcs/services/manager.py
@@ -1,24 +1,24 @@
 from pathlib import Path
-from typing import List, Optional, Tuple
+from typing import List, Optional
 
 from snowflake.cli.api.sql_execution import SqlExecutionMixin
+from snowflake.cli.plugins.spcs.common import strip_empty_lines
+from snowflake.cli.plugins.object.common import Tag
 from snowflake.connector.cursor import SnowflakeCursor
 
-from snowflake.cli.plugins.object.common import Tag
 
 class ServiceManager(SqlExecutionMixin):
-
     def create(
-            self,
-            service_name: str,
-            compute_pool: str,
-            spec_path: Path,
-            num_instances: int,
-            auto_resume: bool,
-            external_access_integrations: Optional[List[str]],
-            query_warehouse: Optional[str],
-            tags: Optional[List[Tag]],
-            comment: Optional[str]
+        self,
+        service_name: str,
+        compute_pool: str,
+        spec_path: Path,
+        num_instances: int,
+        auto_resume: bool,
+        external_access_integrations: Optional[List[str]],
+        query_warehouse: Optional[str],
+        tags: Optional[List[Tag]],
+        comment: Optional[str],
     ) -> SnowflakeCursor:
         spec = self._read_yaml(spec_path)
 
@@ -32,28 +32,27 @@ class ServiceManager(SqlExecutionMixin):
             MIN_INSTANCES = {num_instances}
             MAX_INSTANCES = {num_instances}
             AUTO_RESUME = {auto_resume}
-            """.split("\n")
+            """.splitlines()
 
         if external_access_integrations:
             external_access_integration_list = ",".join(
                 f"{e}" for e in external_access_integrations
             )
-            query.append(f"EXTERNAL_ACCESS_INTEGRATIONS = ({external_access_integration_list})")
+            query.append(
+                f"EXTERNAL_ACCESS_INTEGRATIONS = ({external_access_integration_list})"
+            )
 
         if query_warehouse:
             query.append(f"QUERY_WAREHOUSE = {query_warehouse}")
 
         if tags:
-            tag_list = ",".join(
-                f"{t.name}={t.value_string_literal()}" for t in tags
-            )
+            tag_list = ",".join(f"{t.name}={t.value_string_literal()}" for t in tags)
             query.append(f"TAG ({tag_list})")
 
         if comment:
             query.append(f"COMMENT = {comment}")
 
-        query = "\n".join([q.strip() for q in query if q.strip()])
-        return self._execute_schema_query(query)
+        return self._execute_schema_query(strip_empty_lines(query))
 
     def _read_yaml(self, path: Path) -> str:
         # TODO(aivanou): Add validation towards schema
@@ -71,7 +70,7 @@ class ServiceManager(SqlExecutionMixin):
         )
 
     def logs(
-            self, service_name: str, instance_id: str, container_name: str, num_lines: int
+        self, service_name: str, instance_id: str, container_name: str, num_lines: int
     ):
         return self._execute_schema_query(
             f"call SYSTEM$GET_SERVICE_LOGS('{service_name}', '{instance_id}', '{container_name}', {num_lines});"