+++ tests/spcs/test_services.py
@@ -10,6 +10,7 @@ from tests.testing_utils.fixtures import *
 from snowflake.cli.api.project.util import to_string_literal
 from snowflake.cli.plugins.object.common import Tag
 
+
 @patch(
     "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
 )
@@ -35,7 +36,10 @@ def test_create_service(mock_execute_schema_query, other_directory):
         )
     )
     auto_resume = True
-    external_access_integrations = ["google_apis_access_integration", "salesforce_api_access_integration"]
+    external_access_integrations = [
+        "google_apis_access_integration",
+        "salesforce_api_access_integration",
+    ]
     query_warehouse = "test_warehouse"
     tags = [Tag("test_tag", "test value"), Tag("key", "value")]
     comment = "'user\\'s comment'"
@@ -44,31 +48,37 @@ def test_create_service(mock_execute_schema_query, other_directory):
     mock_execute_schema_query.return_value = cursor
 
     result = ServiceManager().create(
-        service_name, compute_pool, Path(spec_path), num_instances, auto_resume, external_access_integrations,
-        query_warehouse, tags, comment
+        service_name,
+        compute_pool,
+        Path(spec_path),
+        num_instances,
+        auto_resume,
+        external_access_integrations,
+        query_warehouse,
+        tags,
+        comment,
     )
-    expected_query = " ".join([
-        "CREATE SERVICE IF NOT EXISTS test_service",
-        "IN COMPUTE POOL test_pool",
-        'FROM SPECIFICATION $$ {"spec": {"containers": [{"name": "cloudbeaver", "image":',
-        '"/spcs_demos_db/cloudbeaver:23.2.1"}], "endpoints": [{"name": "cloudbeaver",',
-        '"port": 80, "public": true}]}} $$',
-        "WITH MIN_INSTANCES = 42 MAX_INSTANCES = 42",
-        "AUTO_RESUME = True",
-        "EXTERNAL_ACCESS_INTEGRATIONS = (google_apis_access_integration,salesforce_api_access_integration)",
-        "QUERY_WAREHOUSE = test_warehouse",
-        "TAG (test_tag='test value',key='value')",
-        "COMMENT = 'user\\'s comment'",
-    ])
-    actual_query = " ".join(
-        mock_execute_schema_query.mock_calls[0].args[0].split()
+    expected_query = " ".join(
+        [
+            "CREATE SERVICE IF NOT EXISTS test_service",
+            "IN COMPUTE POOL test_pool",
+            'FROM SPECIFICATION $$ {"spec": {"containers": [{"name": "cloudbeaver", "image":',
+            '"/spcs_demos_db/cloudbeaver:23.2.1"}], "endpoints": [{"name": "cloudbeaver",',
+            '"port": 80, "public": true}]}} $$',
+            "WITH MIN_INSTANCES = 42 MAX_INSTANCES = 42",
+            "AUTO_RESUME = True",
+            "EXTERNAL_ACCESS_INTEGRATIONS = (google_apis_access_integration,salesforce_api_access_integration)",
+            "QUERY_WAREHOUSE = test_warehouse",
+            "TAG (test_tag='test value',key='value')",
+            "COMMENT = 'user\\'s comment'",
+        ]
     )
+    actual_query = " ".join(mock_execute_schema_query.mock_calls[0].args[0].split())
     assert expected_query == actual_query
     assert result == cursor
 
-@patch(
-    "snowflake.cli.plugins.containers.services.manager.ServiceManager.create"
-)
+
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager.create")
 def test_create_service_cli_defaults(mock_create, other_directory, runner):
     tmp_dir = Path(other_directory)
     spec_path = tmp_dir / "spec.yml"
@@ -87,12 +97,34 @@ def test_create_service_cli_defaults(mock_create, other_directory, runner):
     """
         )
     )
-    result = runner.invoke(["containers", "service", "create", "--name", "test_service", "--compute-pool", "test_pool", "--spec-path", f"{spec_path}"])
+    result = runner.invoke(
+        [
+            "spcs",
+            "service",
+            "create",
+            "--name",
+            "test_service",
+            "--compute-pool",
+            "test_pool",
+            "--spec-path",
+            f"{spec_path}",
+        ]
+    )
     assert result.exit_code == 0, result.output
-    mock_create.assert_called_once_with(service_name="test_service", compute_pool="test_pool", spec_path=spec_path, num_instances=1, auto_resume=True, external_access_integrations=[], query_warehouse=None, tags=[], comment=None)
-@patch(
-    "snowflake.cli.plugins.containers.services.manager.ServiceManager.create"
-)
+    mock_create.assert_called_once_with(
+        service_name="test_service",
+        compute_pool="test_pool",
+        spec_path=spec_path,
+        num_instances=1,
+        auto_resume=True,
+        external_access_integrations=[],
+        query_warehouse=None,
+        tags=[],
+        comment=None,
+    )
+
+
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager.create")
 def test_create_service_cli(mock_create, other_directory, runner):
     tmp_dir = Path(other_directory)
     spec_path = tmp_dir / "spec.yml"
@@ -112,22 +144,46 @@ def test_create_service_cli(mock_create, other_directory, runner):
         )
     )
     result = runner.invoke(
-        ["containers", "service", "create",
-         "--name", "test_service",
-         "--compute-pool", "test_pool",
-         "--spec-path", f"{spec_path}",
-         "--num-instances", "42",
-         "--no-auto-resume",
-         "--eai-name", "google_api", "--eai-name", "salesforce_api",
-         "--query-warehouse", "test_warehouse",
-         "--tag", "name=value", "--tag", "\"$trange name\"=normal value",
-         "--comment", "this is a test"
-    ])
+        [
+            "spcs",
+            "service",
+            "create",
+            "--name",
+            "test_service",
+            "--compute-pool",
+            "test_pool",
+            "--spec-path",
+            f"{spec_path}",
+            "--num-instances",
+            "42",
+            "--no-auto-resume",
+            "--eai-name",
+            "google_api",
+            "--eai-name",
+            "salesforce_api",
+            "--query-warehouse",
+            "test_warehouse",
+            "--tag",
+            "name=value",
+            "--tag",
+            '"$trange name"=normal value',
+            "--comment",
+            "this is a test",
+        ]
+    )
     assert result.exit_code == 0, result.output
     print(mock_create.mock_calls[0])
-    mock_create.assert_called_once_with(service_name="test_service", compute_pool="test_pool", spec_path=spec_path,
-                                        num_instances=42, auto_resume=False, external_access_integrations=["google_api", "salesforce_api"],
-                                        query_warehouse="test_warehouse", tags=[Tag("name", "value"), Tag("\"$trange name\"", "normal value")], comment=to_string_literal("this is a test"))
+    mock_create.assert_called_once_with(
+        service_name="test_service",
+        compute_pool="test_pool",
+        spec_path=spec_path,
+        num_instances=42,
+        auto_resume=False,
+        external_access_integrations=["google_api", "salesforce_api"],
+        query_warehouse="test_warehouse",
+        tags=[Tag("name", "value"), Tag('"$trange name"', "normal value")],
+        comment=to_string_literal("this is a test"),
+    )
 
 
 @patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._read_yaml")
@@ -142,9 +198,18 @@ def test_create_service_with_invalid_spec(mock_read_yaml):
 
     with pytest.raises(strictyaml.YAMLError):
         ServiceManager().create(
-            service_name, compute_pool, Path(spec_path), num_instances, auto_resume, external_access_integrations,
-            query_warehouse, tags, comment
+            service_name,
+            compute_pool,
+            Path(spec_path),
+            num_instances,
+            auto_resume,
+            external_access_integrations,
+            query_warehouse,
+            tags,
+            comment,
         )
+
+
 @patch(
     "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
 )